requirements.txt
amqp==5.3.1
anyio==4.10.0
argcomplete==3.6.2
arrow==1.3.0
asgiref==3.9.1
backoff==2.2.1
billiard==4.2.1
boto3==1.40.25
botocore==1.40.25
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
composer==0.32.1
coolname==2.2.0
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.6
django-contrib-comments==2.2.0
django-crispy-forms==2.4
django-extensions==4.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
filelock==3.19.1
fsspec==2025.9.0
gql==3.5.3
graphql-core==3.2.6
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mosaicml-cli==0.7.5
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
prompt-toolkit==3.0.36
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytorch-ranger==0.1.1
PyYAML==6.0.2
questionary==2.0.1
requests==2.32.5
rich==13.9.4
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
torch==2.7.0
torch-optimizer==0.3.0
torchmetrics==1.7.4
torchvision==0.22.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
whitenoise==6.9.0
yarl==1.20.1
zipp==3.23.0
